move windows stuff to internal for now, remove --daemon and push to a local registry in e2e test
